+++ octoprint/FS/run_dev.sh
@@ -2,6 +2,8 @@
 set -e
 # #!/usr/bin/with-contenv bashio
 
+
+
 # echo "run.sh"
 echo "run_dev.sh"
 
@@ -24,16 +26,16 @@ create_ingress_user() {
     # bashio::log.info "Create ingress user"
     echo "Create ingress user"
     new_password=`date +%s | sha256sum | base64 | head -c 32 ; echo`
-    octoprint --basedir /config/octoprint/ user add homeassistant --password $new_password --admin # 2> /dev/null
+    octoprint --basedir /config/octoprint user add homeassistant --password $new_password --admin # 2> /dev/null
 }
 
 create_config() {
-    mkdir -p /config/octoprint/
+    mkdir -p /config/octoprint
     if [ ! -f /config/octoprint/config.yaml ]; then
         # bashio::log.info "Create config"
         echo "Create config"
-        mkdir -p /config/octoprint/
-        cd /config/octoprint/
+        mkdir -p /config/octoprint
+        cd /config/octoprint
         touch config.yaml
         echo "accessControl:" >> config.yaml
         echo "  autologinAs: homeassistant" >> config.yaml
@@ -65,7 +67,9 @@ create_config() {
         echo "    debug_logging: false" >> config.yaml
         echo "server:" >> config.yaml
         echo "  commands:" >> config.yaml
-        echo "    serverRestartCommand: supervisorctl reload" >> config.yaml
+        echo "    serverRestartCommand: supervisorctl restart octoprint" >> config.yaml
+        echo "    systemRestartCommand: /restart.sh" >> config.yaml
+        echo "    systemShutdownCommand: /shutdown.sh" >> config.yaml
         echo "webcam:" >> config.yaml
         echo "  ffmpeg: /usr/bin/ffmpeg" >> config.yaml
         # echo "  stream: /webcam/?action=stream" >> config.yaml